% complete math change

+++ src/hpc-server/hpc-bus-service-impl/src/main/java/gov/nih/nci/hpc/bus/impl/HpcDataManagementBusServiceImpl.java
@@ -2145,7 +2145,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
         totalDownloadSize += item.getSize() != null ? item.getSize() : 0;
         totalBytesTransferred +=
             item.getPercentComplete() != null
-                ? item.getPercentComplete() / 100 * item.getSize()
+                ? ((double) item.getPercentComplete() / 100) * item.getSize()
                 : 0;
       }
